requirements.txt
aiohttp==3.7.3
cchardet==2.1.7
cryptography==3.3.1
PyMySQL==0.10.1
PyYAML==5.3.1
SQLAlchemy==1.3.20
GitPython==3.1.11